ai/requirements.txt
google-adk>=0.4.0
chromadb
ragas
datasets
google-generativeai>=0.8.3
google-genai>=0.8.0
langchain-core 
langchain-community
langchain
langchain-google-genai
